+++ Assets/Scripts/Stage/StageInstance.cs
@@ -5,7 +5,7 @@ public sealed class StageInstance
 {
     public StageDto StageDto { get; }
     public int StageIndex => StageDto.index;
-    public int Difficulty => StageDto.difficulty;
+    public double Difficulty => StageDto.difficulty;
     public float SpawnSecond => StageDto.spawnSecond;
     public StageInstance(StageDto stageDto)
     {
